Taskpilot/server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.3
python-dotenv==1.0.1
pydantic==2.7.4
playwright==1.46.0
groq==0.11.0
httpx==0.26.0
